Pull media action functions out to common
